Added a comprehensive Stan model of the data
The `model.py` script loads the data from `crossbot.db`, fits a model to it,
and uses the resulting parameters to generate three plots: `dates.pdf`,
`users.pdf`, and `rates.pdf`. `dates.pdf` shows estimates of date difficulty;
`users.pdf` shows estimates of user skill; and `rates.pdf` shows estimates of
the speed of user improvement over time.

I'm pretty happy with most of the model, and it produces intuitive results, but
I am not yet certain that the model of skill improvement is correct.